completed deviationValues service for reglament table

+++ RUSAL.MetalTapping.DAL/Models/DeviationValues.cs
@@ -4,8 +4,8 @@
     {
         public Guid Id { get; set; }
         public Guid DeviationId { get; set; }
-        public double Value { get; set; }
-        public double CastingRatio { get; set; }
+        public int Value { get; set; }
+        public int CastingRatio { get; set; }
         public Deviation Deviation { get; set; } = null!;
     }
 }
